api/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# HTTP client
httpx==0.27.2
aiohttp==3.10.10

# Content extraction and processing
beautifulsoup4==4.12.3
lxml[html_clean]==5.3.0
html2text==2024.2.26
readability-lxml==0.8.1
python-dateutil==2.9.0

# Data validation
pydantic==2.9.2
pydantic-settings==2.6.0

# Utilities
python-dotenv==1.0.1

# AI/ML (optional - only if OPENAI_API_KEY is set)
openai==1.55.3
numpy==2.2.1
